pyproject.toml
[tool.poetry]
name = "yndx-parse"
version = "0.1.0"
description = ""
authors = ["Alex Ershov <galak654@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
